Mini-batch for reproducing the results
The source code for models/data-preprocessing/training/finetuning/post-hoc is included in this batch. The current scripts are coded with the absolute pathway on the local computer, and the slurm setting is adapted to the environment on the working cluster. Before running the scripts, please change this information based on personal settings. Otherwise, the scripts may fail due to errors.